Cleanup: Remove app dir completely
