+++ lib/ui.rb
@@ -1,0 +1,4 @@
+module Ui
+  module Components
+  end
+end
