+++ slices/discussion/templates/profile/form.rb
@@ -7,7 +7,7 @@ class Discussion::Templates::Profile::Form < Palaver::View
   def template
     div do
       heading2("Create your profile")
-      render Palaver::Components::Form.new(url: "/profile", multipart: true) do
+      render Ui::Components::Form.new(url: "/profile", multipart: true) do
         hidden_field("_csrf_token", csrf_token)
         unless @profile
           field_info { "User name is your unique identifier accross the forum. It has to be unique and cannot be changed later." }
@@ -16,7 +16,7 @@ class Discussion::Templates::Profile::Form < Palaver::View
         input(type: "hidden", value: nil, name: :avatar)
         horizontal_field(label: "Avatar", name: :avatar, type: :file)
         label = @profile ? "Update profile" : "Set up your profile"
-        render Palaver::Components::Form::HorizontalSubmit.new(label:)
+        render Ui::Components::Form::HorizontalSubmit.new(label:)
       end
     end
   end
